+++ CHANGELOG.md
@@ -5,9 +5,16 @@
  https://opensource.org/licenses/MIT
 -->
 # Changelog 
-This file contains changelog of **Arduino Value Storage Library** project
+This file contains changelog of [Arduino Value Storage Library](https://github.com/ru-ace/arduino-value-storage/) project
 
 
-
-## [0.9.0-alpha] - 2018-06-03
+## [Unreleased]
+### Added
+ - Add storage-provider class cavsProviderAT24CX  
+### Changed
+ - Remake storage-config-selector section in avsConfig.h
+## 0.9.0-alpha - 2018-06-03
 - First public release (extracted from inner project)
+
+
+[Unreleased]: https://github.com/ru-ace/arduino-value-storage/compare/v0.9.0-alpha...HEAD